+++ src/render/src/shaders/hips/raytracer/grayscale_to_colormap.frag
@@ -1,9 +1,8 @@
 #version 300 es
 precision mediump float;
-precision lowp sampler2D;
-precision lowp isampler2D;
+precision highp sampler2D;
+precision highp isampler2D;
 precision mediump int;
-
 in vec3 out_vert_pos;
 in vec2 pos_clip;
